Move interface to rest module
