+++ nvi-rest/src/main/java/no/sikt/nva/nvi/rest/create/CreateNoteHandler.java
@@ -8,7 +8,7 @@ import com.amazonaws.services.lambda.runtime.Context;
 import java.net.HttpURLConnection;
 import java.util.Objects;
 import java.util.UUID;
-import no.sikt.nva.nvi.common.ViewingScopeHandler;
+import no.sikt.nva.nvi.rest.ViewingScopeHandler;
 import no.sikt.nva.nvi.common.client.OrganizationRetriever;
 import no.sikt.nva.nvi.common.db.CandidateRepository;
 import no.sikt.nva.nvi.common.db.PeriodRepository;
