+++ nvi-rest/src/main/java/no/sikt/nva/nvi/rest/remove/RemoveNoteHandler.java
@@ -5,7 +5,7 @@ import static nva.commons.core.attempt.Try.attempt;
 import com.amazonaws.services.lambda.runtime.Context;
 import java.net.HttpURLConnection;
 import java.util.UUID;
-import no.sikt.nva.nvi.common.ViewingScopeHandler;
+import no.sikt.nva.nvi.rest.ViewingScopeHandler;
 import no.sikt.nva.nvi.common.client.OrganizationRetriever;
 import no.sikt.nva.nvi.common.db.CandidateRepository;
 import no.sikt.nva.nvi.common.db.DynamoRepository;
